13 Tests passing

+++ car.h
@@ -26,6 +26,6 @@ double get_fill_level(Car car);
 double get_acceleration_rate(Car car);
 int get_speed(Car car);
 void init();
-void set_acceleration_rate(Car car, int acceleration);
+void set_acceleration_rate(Car car, double acceleration);
 void accelerate(Car car);
 #endif
